refactor: renamed mdx file from server to file
